refactor: centralize formatting helpers
